pyproject.toml
[tool.poetry]
name = "airflow-emr-serverless-pyspark-demo"
version = "0.1.0"
description = ""
authors = ["Nhat Minh DOAN"]

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.26.142"
apache-airflow = {extras = ["amazon"], version = "^2.6.1"}

[tool.poetry.dev-dependencies]
black = "^22.3.0" # The uncompromising code formatter.
isort = "^5.9.3" #A Python utility / library to sort Python imports.
# https://flakehell.readthedocs.io/
flakehell = "0.9.0"
flake8 = "^3.9.2"
# https://github.com/DmytroLitvinov/awesome-flake8-extensions
flake8-bugbear = "^20.11.1"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.3.1"
flake8-docstrings = "^1.5.0"
flake8-eradicate = "^1.0.0"
flake8-isort = "^4.0.0"
flake8-mutable = "^1.2.0"
flake8-pytest-style = "^1.3.0"
# flake8-spellcheck = "^0.20.0"
pep8-naming = "^0.11.1"
pytest = "^7.3.1"
pytest-cov = "^2" # Pytest plugin for measuring coverage.
pytest-mock = "^3"
xdoctest = "^0.15.0" # A rewrite of the builtin doctest module
coverage = { version = "*", extras = ["toml"] }

# see https://black.readthedocs.io/en/stable/
[tool.black]
target-version = ["py38"]
skip-string-normalization = true
line_length = 120

# see https://timothycrosley.github.io/isort/
[tool.isort]
combine_as_imports = true
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
force_grid_wrap = 0
ensure_newline_before_comments = true
line_length = 120

[tool.flakehell]
extended_default_ignore = []
exclude = ["README.md"]
format = "colored"
max_line_length = 120
show_source = true
max-complexity = 18
docstring-convention = "google"
#whitelist = "../../allowlist.txt"

[tool.flakehell.plugins]
#flake8-bandit = ["+*", "-S322"]
flake8-bugbear = ["+*"]
flake8-builtins = ["+*"]
flake8-comprehensions = ["+*"]
#flake8-darglint = ["+*", "-DAR201"]
# explicitly disable docstring
# D100: Missing docstring in public module
# D101: Missing docstring in public class
# D102: Missing docstring in public method
# D103: Missing docstring in public function
# D104: Missing docstring in public package
# D105: Missing docstring in magic method
# D107: Missing docstring in __init__
# D202: No blank lines allowed after function docstring
flake8-docstrings = ["+*", "-D100", "-D101", "-D102", "-D103", "-D104", "-D105", "-D107", "-D202" ]
flake8-eradicate = ["+*", "-E800"]
flake8-isort = ["+*"]
flake8-mutable = ["+*"]
flake8-pytest-style = ["+*"]
# flake8-spellcheck = ["+*", "-SC200", "-SC100"]
mccabe = ["+*"]
pep8-naming = ["+*"]
pycodestyle = ["+*", "-W503", "-F841", "-E203"]
pyflakes = ["+*", "-W503", "-F821", "-F841"]
pylint = ["+*", "-F841"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers --pdbcls=tests:Debugger -r sxX --cov=some_dags --cov-report=html --cov-report=term-missing:skip-covered"


# see https://coverage.readthedocs.io/en/coverage-5.0.3/config.html
[tool.coverage.paths]
source = ["some_dags"]

[tool.coverage.run]
branch = true
source = ["some_dags"]
omit = ["tests/*", ".venv/*", "*/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError"]


[build-system]
requires = ["poetry>=1.1.12"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pypi-public"
url = "https://pypi.org/simple/"
